Update README.md
1 Introduction
This project will feel somewhat familiar in that it is nearly identical to the preceding project: there is a coding problem and a puzzle-solving problem. The major change is that everything is at the assembly level:
• Problem 1 re-works the Thermometer functions from the previous project in x86-64 Assembly rather than C
• Problem 2 involves analyzing a binary executable to provide it with the correct input to “defuse” the executable much like the previous project’s Puzzlebox problem
Working with assembly will get you a much more acquainted with the low-level details of the x86-64 platform and give you a greater appreciation for “high-level” languages (like C).
2 Download Code and Setup
Download the code pack linked at the top of the page. Unzip this which will create a project folder. Create new files in this folder. Ultimately you will re-zip this folder to submit it.